refactor(refs T30969): move DpSlidingPagination to components root
